requirements.txt
ecdsa==0.19.0
pycryptodomex==3.21.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
tqdm==4.66.4
xxhash==3.5.0
pytest==7.4.3
numpy==1.26.4
# GPU acceleration (optional; install separately if CUDA available)
# cupy-cuda12x==13.0.0  # Replace 12x with your CUDA version (11, 12, etc.)
